backend/requirements.txt
flask==3.0.3
faster-whisper==1.0.3
numpy==1.26.4
pytest==8.3.2
pyannote.audio==3.3.1
torch==2.4.1
torchaudio==2.4.1
python-dotenv==1.0.1
